+++ cli/Fixtures/ios_app_with_framework_headers_in_buildable_folders/App/Sources/ContentView.swift
@@ -1,0 +1,18 @@
+import SwiftUI
+
+public struct ContentView: View {
+    private let greeting = UseFrameworkGreeting() as String
+
+    public init() {}
+
+    public var body: some View {
+        Text(greeting)
+            .padding()
+    }
+}
+
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        ContentView()
+    }
+}