+++ cli/Fixtures/ios_app_with_framework_headers_in_buildable_folders/Framework/Sources/Framework.h
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+#import <Framework/bar.h>
+#import <Framework/baz.h>
+
+FOUNDATION_EXPORT NSString *FrameworkGreeting(void);